Fix extension not working if the header is not loaded
We used to start moving things as soon as DOMContentLoaded happened but this means that just the basic document is
ready not really the entire webapp.

All the work is done inside the header icons so we wait until the element with class drawer__header appears and then
assume is safe to proceed.