Improve scrolling focus behaviour
Rather than going directly into the viewport we just scroll the bare minimum to show the remaining of the article
that was hidden. If we are already at the very top we force the scrollable to go down a single pixel in order to
guarantee that the streaming doesn't push the timeline anymore.