Drop target of reply rather than waiting for it
We were waiting until the content was drawn and matching the content of the article before trying to scroll to make
sure we had the correct height but that doesn't make much sense since the composer is appended below the status and
we can already easily read it.

This also means we no longer have to hook the cancel button because it won't even be clickable.