Workaround "dismiss changes" nag
Recent versions of Mastodon introduced a nag to alert the user about starting a new a
reply whe some other is already in progress.

This breaks to a heavy extent the navigation of the interface when using the extension so
we must avoid it altogether by always cleaning the textarea before attempting to move it
to the new position.