Let the user interact with the discard too nag
If we are already replying to something or the textarea is not empty we just wait for the nag to appear and then hook the REPLY button to move the composer only if we decided to overwrite.
In the case of using the reply hotkey "R" we need to test it before releasing the key because mastodon starts working on keydown.